Create FileSystemTargetTestMixin
Ensure that all FileSystemTarget satisfy a
minimal common interface
