+++ test/requirements.txt
@@ -1,5 +1,6 @@
 mock
 moto
+HTTPretty>=0.8.1
 nose
 psycopg2
 unittest2
